Fixing error

+++ src/main/java/id/ac/ui/cs/advprog/yomu/dto/CompletedQuizRequest.java
@@ -7,6 +7,6 @@ import lombok.Setter;
 @Getter
 public class CompletedQuizRequest {
   private int score;
-  private double accuracy;
+  private int accuracy;
 
 }
