+++ src/main/java/id/ac/ui/cs/advprog/yomu/dto/QuizSubmitResponse.java
@@ -11,7 +11,7 @@ import java.util.Map;
 @Builder
 public class QuizSubmitResponse {
   private int score;
-  private double accuracy;
+  private int accuracy;
   private int totalQuestions;
   private int correctAnswers;
   private double timeTaken;
